Product update actions

+++ models/src/main/java/io/sphere/sdk/products/commands/updateactions/RemovePrice.java
@@ -5,7 +5,7 @@ import io.sphere.sdk.products.ProductUpdateScope;
 
 /**
  *
- * {@include.example io.sphere.sdk.products.ProductCrudIntegrationTest#removePriceUpdateAction()}
+ * {@include.example io.sphere.sdk.products.commands.ProductUpdateCommandTest#removePrice()}
  */
 public class RemovePrice extends StageableProductUpdateAction {
     private final int variantId;
@@ -17,7 +17,7 @@ public class RemovePrice extends StageableProductUpdateAction {
         this.price = price;
     }
 
-    public long getVariantId() {
+    public int getVariantId() {
         return variantId;
     }
